# docs/swagger.yaml
basePath: /api/v1
definitions:
  ans-spareparts-api_internal_features_category.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  ans-spareparts-api_internal_features_inventory.InventoryResponse:
    properties:
      id:
        type: integer
      productID:
        type: integer
      quantity:
        type: integer
    type: object
  ans-spareparts-api_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_features_product.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
    type: object
  internal_features_product.LiteProductResponse:
    properties:
      categoryID:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        format: float64
        type: number
      sku:
        type: string
    type: object
  internal_features_product.ProductDetailResponse:
    properties:
      category:
        $ref: '#/definitions/ans-spareparts-api_internal_features_category.CategoryResponse'
      categoryID:
        type: integer
      description:
        type: string
      id:
        type: integer
      inventory:
        $ref: '#/definitions/ans-spareparts-api_internal_features_inventory.InventoryResponse'
      name:
        type: string
      price:
        format: float64
        type: number
      sku:
        type: string
    type: object
  internal_features_product.ProductListResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/internal_features_product.LiteProductResponse'
        type: array
      total:
        format: int64
        type: integer
    type: object
  internal_features_product.UpdateProductRequest:
    properties:
      categoryID:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        format: float64
        type: number
      sku:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: suppourt@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Point of sale system API built with Clean Architechture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0html
  termsOfService: http://swagger.io/terms/
  title: ANS Sparepart API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Getall products with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_product.ProductListResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product creation request
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_features_product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_features_product.ProductDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a new product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product (admin only)
      parameters:
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get product details by product ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_product.ProductDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get product by ID
    put:
      consumes:
      - application/json
      description: Update product details (admin/manager only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_features_product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_product.ProductDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /users/profile:
    delete:
      consumes:
      - application/json
      description: Delete current authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ans-spareparts-api_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - User Profile
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization Type "Bearer" followed by a space and JWT token.
    type: apiKey
swagger: "2.0"